chore: move attendanceExists and ApiExample
