+++ src/widgets/BhandaraCheckin/api/attendanceExitsts.ts
@@ -1,6 +1,6 @@
-import { AttendanceExistsUser } from "../../types";
-import { fetchWithToken } from "../utils";
-import { API_URL, API_VERSION, EVENT_ID } from "./env-variables";
+import { AttendanceExistsUser } from "../types";
+import { fetchWithToken } from "./utils";
+import { API_URL, API_VERSION, EVENT_ID } from "./server-apis/env-variables";
 
 export const attendanceExists = (
   user: AttendanceExistsUser
