+++ src/widgets/BhandaraCheckin/api/server-apis/index.ts
@@ -1,4 +1,3 @@
-export * from "./attendanceExitsts";
 export * from "./postAttendance";
 export * from "./searchUser";
 export * from "./login";
